==> notebooks/01_preprocessing.ipynb <==
import pandas as pd

walmart = pd.read_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/processed/walmart.csv')

# Concatenate the address fields
walmart['Address'] = walmart['Address'] + ', ' + walmart['City'] + ', ' + walmart['State']

walmart

walmart = walmart.drop(walmart.columns[[0, 1]], axis=1)

# Save the modified DataFrame to a CSV file
walmart.to_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/processed/walmart.csv', index=False)
