==> notebooks/14_feature.ipynb <==
import os
import pandas as pd
import requests
from dotenv import load_dotenv
from urllib.parse import urlencode

# Load environment variables
load_dotenv()

# API key from .env file
API_KEY = os.getenv('DISTANCE_MATRIX_API_KEY')

# Load the geocoded data
walmart_df = pd.read_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/interim/walmart_sample_geocoded.csv')
zips_df = pd.read_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/interim/zips_sample_geocoded.csv')

# Extract the state from the Walmart addresses and add it as a new column
walmart_df['State'] = walmart_df['Address'].apply(lambda x: x.split(',')[-1].strip())

# Output Files (CSV)
output_file = '/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/interim/distance_sample.csv'

walmart_df

zips_df

def fetch_and_append_distances(state, output_file):
    # Filter dataframes for the state
    state_walmart_df = walmart_df[walmart_df['State'] == state]
    state_zips_df = zips_df[zips_df['state_name'] == state]

    for _, zip_row in state_zips_df.iterrows():
        origin = f"{zip_row.latitude},{zip_row.longitude}"
        for _, walmart_row in state_walmart_df.iterrows():
            destination = f"{walmart_row.latitude},{walmart_row.longitude}"
            
            # API request
            url = f"https://api.distancematrix.ai/maps/api/distancematrix/json?origins={origin}&destinations={destination}&key={API_KEY}"
            response = requests.get(url)
            
            # Initialize the dictionary that will hold our row data
            row_data = {
                'origin': origin,
                'destination': destination,
                'distance_in_time': 'N/A'  # Default value in case of an error
            }

            if response.status_code == 200:
                data = response.json()
                if data['status'] == 'OK':
                    distance_info = data['rows'][0]['elements'][0]
                    if distance_info['status'] == 'OK':
                        row_data['distance_in_time'] = distance_info['duration']['text']
                        
            # Append the result to the CSV file
            with open(output_file, 'a') as f:
                pd.DataFrame([row_data]).to_csv(f, header=False, index=False)

# Iterate through each state and append results to the CSV
for state in zips_df['state_name'].unique():
    fetch_and_append_distances(state, output_file)

print("Distance data appended to CSV successfully.")
