==> notebooks/19_feature.ipynb <==
import pandas as pd

# Load the time_distance.csv file
time_distance = pd.read_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/processed/time_distance.csv')

# Initialize a counter
total_combinations = 0

# Loop through each row and count the number of Walmart stores for each ZIP code
for index, row in time_distance.iterrows():
    # Split the string of Walmart store numbers and count them
    stores = str(row['walmarts_pot_10']).split(' ')
    # Filter out any empty strings in case of missing data
    stores = [store for store in stores if store.strip()]
    total_combinations += len(stores)

print(f"Total number of combinations: {total_combinations}")


import math

# Population size
N = total_combinations

# Confidence Level (95%) Z-Score
Z = 2.576

# Margin of error (e.g., 5%)
e = 0.02

# Assuming p = 0.5 (maximum variability)
p = 0.5

# Calculate the sample size
sample_size = ((Z**2 * p * (1 - p)) / e**2) / (1 + ((Z**2 * p * (1 - p)) / (e**2 * N)))
sample_size = math.ceil(sample_size)

sample_size
