==> notebooks/25_model.ipynb <==
import pandas as pd
import numpy as np

# Load the population dataset
population_data = pd.read_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/processed/final_cleaned_time_distance.csv')

# Coefficients and intercepts for the models
coefficients_distance = {'great_circle_distance': 1.335168, 'ruca_Rural': 0.017702, 'ruca_Suburban': -0.144439, 'ruca_Urban': 0.126737}
intercept_distance = 0.34598873463459157

# Assuming you have coefficients and intercept for the driving_time model
coefficients_time = {'great_circle_distance': 1.809767, 'ruca_Rural': -1.119989, 'ruca_Suburban': -0.729187, 'ruca_Urban': 1.849177}  # Replace with actual values
intercept_time = 2.961072620208963  # Replace with actual value

# Prediction function for the linear regression model
def predict_linear_regression(row, coefficients, intercept):
    return sum(row[feature] * coef for feature, coef in coefficients.items()) + intercept

# Predict driving_distance
population_data['driving_distance'] = population_data.apply(predict_linear_regression, axis=1, coefficients=coefficients_distance, intercept=intercept_distance)

# Predict driving_time
population_data['driving_time'] = population_data.apply(predict_linear_regression, axis=1, coefficients=coefficients_time, intercept=intercept_time)

# Binary classification for within_10_miles
CONSERVATIVE_THRESHOLD = 8  # miles
population_data['within_10_miles'] = np.where(population_data['driving_distance'] <= (CONSERVATIVE_THRESHOLD * 1.60934), 1, 0)

# Save the updated DataFrame
population_data.to_csv('/Users/pintoza/Desktop/dev/data-science/walmart-proximity/data/processed/final_population_predictions.csv', index=False)
